\name{CaseY.Database}
\alias{CaseY.Database}
\docType{data}
\title{
Allele counts for second example of
(Table) 3 of Balding and Buckleton, 2010
}
\description{
The marker allele and counts are given in a data frame.
}
\usage{data(Bates.Database)}
\format{
  A data frame with 104 observations on the following 3 variables.
  \describe{
    \item{\code{marker}}{Marker name}
    \item{\code{allele}}{a numeric vector}
    \item{\code{count}}{a numeric vector}
  }
}

\references{
Balding DJ, Buckleton J, Interpreting low template DNA profiles,
Forensic Science International: Genetics, 4: 1-10, 2009, doi: 10.1016/j.fsigen.2009.03.003.
http://www.zebfontaine.eclipse.co.uk/djb.htm
}
\examples{
data(Bates.Database)
}
\keyword{datasets}
