\encoding{UTF-8}
\name{forensim-package}
\alias{forensim-package}
\alias{forensim}
\title{The forensim package}

\description{forensim is dedicated to the interpretation of forensic DNA mixtures through statistical methods.
It relies on three S4 classes that facilitate the  manipulation and the storage of genetic data produced in 
forensic casework: \linkS4class{tabfreq},   \linkS4class{simugeno} and \linkS4class{simumix}.\cr
\linkS4class{tabfreq} objects are used to store allele frequencies, \linkS4class{simugeno} objects are 
used to store genotypes and 
\linkS4class{simumix} objects are used to store DNA mixtures.\cr
For more information about these classes type 'class ?tabfreq', 'class ?simugeno' and 'class ?simumix'.\cr


 }

\author{Hinda Haned  <h.haned@nfi.minjus.nl>}

\keyword{manip}
\keyword{datagen}


