\encoding{UTF-8}

\name{Hbsimu}
\alias{Hbsimu}

\title{
A Tcl/Tk simulator of the  heterozygous balance
}
\description{
\code{Hbsimu}  is a user-friendly graphical interface simulating the
     heterozygous balance of heterozygous profiles generated according
     to the simulation model described in Gill et al. (2005)

}
\usage{
Hbsimu()
}


\author{ Hinda Haned  \email{contact@hindahaned.info} }

\references{
Gill P, Curran J and Elliot K. A graphical simulation model of the entire DNA process associated with the analysis of short 
tandem repeat loci.\emph{ Nucleic Acids Research} 2005, 33(2): 632-643.
}
\examples{
Hbsimu()
}

\keyword{datagen}
