\name{LRmixTK}
\alias{LRmixTK}
\title{GUI for the LR function
}
\description{
User-friendly graphical user interface for the LR calculator LR. }
\usage{
LRmixTK(verbose)
}
\arguments{
  \item{verbose}{if TRUE, progress is written to the console}
}
\value{No return value, called to show GUI.}

\author{Hinda Haned \email{contact@hindahaned.info}}
\keyword{htest}
