\name{ngm}
\alias{ngm}
\docType{data}
\title{
Allele frequencies for the new generation markers NGM, for the Caucasian US population}

\description{Allele frequencies for 15 autosomal short tandem repeats  loci in the  American Caucasian population. }

\usage{data(ngm)}



\references{Budowle, B.; Ge, J.; Chakraborty, R.; Eisenberg, A.; Green, R.; Mulero, J.; Lagace, R. & Hennessy, L. Population genetic analyses of the NGM STR loci International Journal of Legal Medicine, 2011, 1-9
}
\examples{
library(forensim)
data(ngm)
boxplot(ngm$tab)
}
\keyword{datasets}

