\name{sgmNorway}
\alias{sgmNorway}
\docType{data}
\title{
Allele frequencies for the new generation markers sgm, for the Norwegian population}

\description{Allele frequencies for 10 autosomal short tandem repeats  loci in the Norwegian population.}

\usage{data(sgmNorway)}

\references{Andreassen, R., S. Jakobsen, and  Mevaag, B., Norwegian population data for the 10 autosomal STR loci in the AMPFlSTR(R) SGM Plus(TM) system. Forensic Science International, 2007. 170(1): p. 59-61.
 
}

\keyword{datasets}

