\encoding{UTF-8}
\name{simPCR2TK}
\alias{simPCR2TK}
\title{
A Tcl/Tk graphical interface for the polymorphism chain reaction simulation model
}
\description{
\code{simPCR2TK} is a user-friendly graphical interface for the \code{simPCR2} function that implements a simulation model
for the polymorphism chain reaction.
}
\usage{
simPCR2TK()
}


\references{
Gill P, Curran J and Elliot K. A graphical simulation model of the entire DNA process associated with the analysis of short 
tandem repeat loci.\emph{ Nucleic Acids Research} 2005, 33(2): 632-643.
}

\seealso{ \code{\link{simPCR2}}}  

\author{ Hinda Haned  \email{haned@biomserv.univ-lyon1.fr} }



\examples{

#launch the graphical interface
simPCR2TK()

}
%\keyword{manip}
\keyword{datagen}

