\name{strveneto}
\Rdversion{1.1}
\alias{strveneto}
\docType{data}
\title{
	Population study of three miniSTR loci in Veneto (Italy)
}
\description{
Allele frequencies for three short tandem repeats loci D10S1248, D2S441 and D22S1045 in a sample of 198 individuals born in Veneto,
 Italy. These loci are commonly used in forensic DNA characterization.

}
\usage{data(strveneto)}
\format{\code{strveneto} is a tabfreq object}

\references{
Turrina S, Atzei R, De Leo D. Population study of three miniSTR loci in Veneto (Italy). Forensic Sci Int Genetics 2008;
1(1);378-379
}

\examples{
data(strveneto)
#allele frequencies
strveneto@tab

}
\keyword{datasets}
