\encoding{UTF-8}
\name{virtualClasses}
\docType{class}
\alias{vectorOrNULL-class}
\alias{vectorOrdataframe-class}
\alias{matrixOrdataframe-class}
\alias{factorOrNULL-class}
\alias{listOrdataframe-class}
\alias{characterOrNULL-class}
\title{Virtual classes for forensim}
\description{Virtual classes that are only for internal use in forensim}
\section{Objects from the Class}{A virtual Class: programming tool, not intended for objects creation.}
\author{Hinda Haned \email{haned@biomserv.univ-lyon1.fr} }
\keyword{classes}