\encoding{UTF-8}
\name{forensim-package}
\alias{forensim-package}
\alias{forensim}
\docType{package}
\title{The forensim package}

\description{forensim is dedicated to the interpretation of forensic DNA mixtures through statistical methods.
It relies on three S4 classes that facilitate the  manipulation and the storage of genetic data produced in 
forensic casework: \linkS4class{tabfreq},   \linkS4class{simugeno} and \linkS4class{simumix}.\cr
\linkS4class{tabfreq} objects are used to store allele frequencies, \linkS4class{simugeno} objects are used to store genotypes and 
\linkS4class{simumix} objects are used to store DNA mixtures.\cr
For more informations about these classes type 'class ?tabfreq', 'class ?simugeno' and 'class ?simumix'.\cr

%Statistical tools for interpretation of the functions
%-\code{\link{mincontri}}: the maximum allele count method
%-\code{\link{likestim}}: the maximum of likelihood method

 }

\author{Hinda Haned  <haned@biomserv.univ-lyon1.fr>}

\keyword{manip}
\keyword{datagen}


