\encoding{UTF-8}
\name{forensim-package}
\alias{forensim-package}
\alias{forensim}
\docType{package}
\title{The forensim package}

\description{forensim is dedicated to the interpretation of forensic DNA mixtures through statistical methods.
It provides  three S4 classes that facilitate the  manipulation and the storage of genetic data produced in forensic casework:\cr
 - \linkS4class{tabfreq} objects are used to store allele frequencies, from either one or several populations \cr
 - \linkS4class{simugeno} objects are used to store genotypes \cr
 - \linkS4class{simumix} objects are used to store DNA mixtures \cr
%forensim also contains simulation tools that allow the study of the performances of the most widely used  statistical methods  for forensic DNA 
%interpretation, based on qualitative data:\cr

%Statistical tools for interpretation of the functions
%-\code{\link{mincontri}}: the maximum allele count method
%-\code{\link{likestim}}: the maximum of likelihood method

 }

\author{Hinda Haned  <haned@biomserv.univ-lyon1.fr>}

\keyword{manip}
\keyword{datagen}


